+++ manuals/stx_manual_ai/book.py
@@ -98,6 +98,7 @@ st_book([
     blocks.bck_ai_image_overview,
     blocks.bck_ai_image_usage,
     blocks.bck_ai_image_example,
+    blocks.bck_ai_image_editor,
     blocks.bck_memory,
     blocks.bck_stx_guide,
     blocks.bck_custom_commands,